+++ udigest/tests/deterministic_hash.rs
@@ -22,13 +22,13 @@ fn sha2_256() {
     let alice_hash = udigest::hash::<sha2::Sha256>(&ALICE);
     assert_eq!(
         hex::encode(alice_hash.as_slice()),
-        "99e258d6a6ccc430a50dcbf4e9c8cfb59ad0b94b96b83f0182a9a68eb1c5438f",
+        "49c43095eaffb3e3232dd23940686d3c6fb80e5ff82b5a09d336ad32369ca9df",
     );
 
     let bob_hash = udigest::hash::<sha2::Sha256>(&BOB);
     assert_eq!(
         hex::encode(bob_hash.as_slice()),
-        "28474b5dec79b222b74badc2d78f9f81c0fbfd1ee04a134947cd07f44237ade3",
+        "3537c188336cb93f58df79149fb035fd132f23fac58a6e94d014178aeaa1c88e",
     );
 }
 
@@ -41,20 +41,20 @@ fn shake256() {
     alice_hash_reader.read(&mut hash);
     assert_eq!(
         hex::encode(&hash),
-        "54809cf7b06438f9508785fb5e46bdfd7714b39b026e86fa7cc8a8442ae10bd5\
-        49baeced19ff0642b042ae4e92636536baec5748dad99e71fc53a4361734973ae\
-        2c4f1547305a76addd5b6076509ddbf91bd5beb71ba09598e265704d1e9a1c0c3\
-        5fae7f8e4958ceb38962fc8e6fc56e32bef4e88f64bc8a88f88a"
+        "ee629bcc426422887fe6f9a9a3384128bd5efc3c623a4599c8526c24a97972be\
+        2a325ef03c95ac649b77f0193c901c942762e93fd939372ef484681220c6fc0b\
+        0dc12be8c6b9ee914dac34697d0deeb3a3e510f24a1b0bfc24d144b639a66c6a\
+        4c5a772b178eed159f87b581bb49aafdcdcca525fd57749aab6c32"
     );
 
     let mut bob_hash_reader = udigest::hash_xof::<sha3::Shake256>(&BOB);
     bob_hash_reader.read(&mut hash);
     assert_eq!(
         hex::encode(&hash),
-        "f68ca9eeb7e09657fc54a5cbbd50acdd6d9fccd29ec1a3eb460b673ea59d64a9\
-        b2ec8be97c7d7858ad6724cf8c27299569bd72193c77bb339883214a4477c0762\
-        f9cf31a2d698562f57dff5ede03d6928feba694975445e7dabe3d67e67b710f26\
-        11f4f14471917bd447d199c32eb93dbcaf1fdbefe05132911991"
+        "56cd71e796fc94176923b73bfe3f659ea7a9a666a2faae6020d1c4f41a51035a\
+        e7965583087f1badf452a40036499d54075350d8e64e5b68b0f3f52c286c15e3\
+        cb010249754a0c7f263d14c7a284da134ca133df84c62d80adfdb0ec0d5c3f0a\
+        50e479dd025b27fb875c34ba72d9abc7a5990ce8c7f3c282dd6a0c"
     );
 }
 
@@ -65,14 +65,14 @@ fn blake2b() {
     udigest::hash_vof::<blake2::Blake2bVar>(&ALICE, &mut out).unwrap();
     assert_eq!(
         hex::encode(&out),
-        "91d1ce144fd46ed5400895c8db5f2b39c95870020c6627af034a9fa09c2f2cc3\
-        f4c8c7d4e8d38ff16e4f54360b4387c0439cf30c51c21c78f904cda9205023"
+        "57b2a8a078ca3b04dc72b308696bc4715c62593b461608bff01388ef3bd49fed\
+        244bd2e9407965ec2bfe13781ae3cd28ea0cb08fb4b46824ea7909c488fec8"
     );
 
     udigest::hash_vof::<blake2::Blake2bVar>(&BOB, &mut out).unwrap();
     assert_eq!(
         hex::encode(&out),
-        "2f916c687c82c0f37d31df061c0453e98d0655e1877d4a55ec1507514822a2c4\
-        b7cac3ca66a5e3deb678f915210e93f2fc14591b987f121083623ab024ece4"
+        "83aa6240d105ec1b496e6963dbab3e48fd09860b734c963b59ee764781d922f1\
+        207405232c1d84965b32f6a73b182b224d1533859f586c332377fe4a39489e"
     );
 }